[refactoring] login
